requirements.txt
pymongo 
google-generativeai 
pandas 
python-dotenv 
matplotlib 
seaborn
Pillow
uvicorn
jinja2